Added HTML-boilerplatte
This is an example of a commit message using the 7 best practices
or whatever. I am stopping each line at 50 characters, the soft
limit. The hard limit is 72 characters. The first line is by itself
and it is a "imperative" line, meaning it states a task.

These new files are the last addition to the basic training that
included creating images. Link below:
https://www.theodinproject.com/paths/foundations/courses/
foundations/lessons/commit-messages